Changed the font size type from int to float.

+++ cat-boot-report-pdf/src/main/java/cc/catalysts/boot/report/pdf/config/PdfTextStyle.java
@@ -8,12 +8,12 @@ import java.awt.*;
 public class PdfTextStyle {
 
 
-    private int fontSize;
+    private float fontSize;
     private PdfFont font;
     private Color color;
     private String style;
 
-    public PdfTextStyle(int fontSize, PdfFont defaultFont, Color color, String style) {
+    public PdfTextStyle(float fontSize, PdfFont defaultFont, Color color, String style) {
         this.fontSize = fontSize;
         this.font = defaultFont;
         this.color = color;
@@ -34,7 +34,7 @@ public class PdfTextStyle {
         color = new Color(Integer.valueOf(split[2].substring(1), 16));
     }
 
-    public int getFontSize() {
+    public float getFontSize() {
         return fontSize;
     }
